sr(api): move find-paginated-training in devcomp
